+++ src/lib/temple/front/index.ts
@@ -1,4 +1,5 @@
 export * from 'lib/temple/types';
+export * from 'lib/temple/analytics-types';
 export * from 'lib/temple/helpers';
 export * from 'lib/temple/assets';
 export * from 'lib/temple/metadata';
