+++ src/lib/temple/types.ts
@@ -1,7 +1,12 @@
 import { Estimate } from '@taquito/taquito';
 import { TempleDAppMetadata, TempleDAppNetwork } from '@temple-wallet/dapp/dist/types';
 
-import { AnalyticsEventCategory } from '../analytics';
+import {
+  TempleSendPageEventRequest,
+  TempleSendPageEventResponse,
+  TempleSendTrackEventRequest,
+  TempleSendTrackEventResponse
+} from './analytics-types';
 
 type NonEmptyArray<T> = [T, ...T[]];
 
@@ -648,32 +653,6 @@ export interface TempleRemoveDAppSessionRequest extends TempleMessageBase {
 export interface TempleRemoveDAppSessionResponse extends TempleMessageBase {
   type: TempleMessageType.DAppRemoveSessionResponse;
   sessions: TempleDAppSessions;
-}
-
-export interface TempleSendTrackEventRequest extends TempleMessageBase {
-  type: TempleMessageType.SendTrackEventRequest;
-  userId: string;
-  rpc: string | undefined;
-  event: string;
-  category: AnalyticsEventCategory;
-  properties?: object;
-}
-
-export interface TempleSendTrackEventResponse extends TempleMessageBase {
-  type: TempleMessageType.SendTrackEventResponse;
-}
-
-export interface TempleSendPageEventRequest extends TempleMessageBase {
-  type: TempleMessageType.SendPageEventRequest;
-  userId: string;
-  rpc: string | undefined;
-  path: string;
-  search: string;
-  additionalProperties: object;
-}
-
-export interface TempleSendPageEventResponse extends TempleMessageBase {
-  type: TempleMessageType.SendPageEventResponse;
 }
 
 export type OperationsPreview = any[] | { branch: string; contents: any[] };
